ipmitool: Ignore ErrChannelCipherSuites errors
Newer versions of ipmitool try to query the BMC for supported ciphers,
if it encounters an error a message will be emitted to stderr. PBnJ will
interpret a line on stderr as a hard error, even though ipmitool will
continue running the command. So we ignore any cipher error messages.