add configurable cipher
some older bmc models don't have support for newer ciphers that are
supported by ipmitool. This change allows for a specific cipher to
be defined by default, as well as adds an api header to allow the
cipher to be defined at a per request level to support older hardware.