Implemented power actions:
Add functionality to power a machine on, off, reset, hardoff,
cycle, and get power status.

All requests will try connecting to a given BMC in the order
in which connections are placed in the slice of `powerConnections`
in machine.go#63. bmclib,  ipmi, and redfish. If none are able
to connect then an error is returned. For all successful
connections, it will try to perform the power action.
If one completes successfully then no others are tried.
If all fail, then an error is returned.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>