Decrease BMC actions timeout:
Current 5 minute timeout creates a undesired user experience
as the user has to wait up to 5 minutes for a task to finish
and the task status to be updated as complete. Changing the
timeout to maybe lower than it should be but this will
improve the user experience, for now.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>