Return actual gRPC error from TaskService.Status:
Not returning an actual gRPC error causes the log line
thats returned for each call to an RPC to be of error
level and contain a stacktrace. Using an actual gRPC
error in the return keeps the log level at info
and just returns the contents of the error. Better
user experience when searching and filtering logs.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>