Create equinix metal specific functionality:
the previous http version of pbnj defined a power-cycle to
do the following:
1. if a machine is currently powered on, power it off
and then power it on again.
OR
2. if a machine is currently powered off, power it on

This definition does not reflect how any other BMC tool
in the industry defines power-cycle. This definition is
used by equinix metal pbnj clients and these same
clients did not change their definition when migrating
to pbnj.next :(

This change is being created in a branch of its own
that should NEVER be merged to master/main.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>