Upgrade bmclib library v0.4.10 -> v0.4.13:
This is wanted because v0.4.13 has the ability to
return metadata about which providers were attempted
successfully and unsuccessfully. More commits to come
to enable retrieving and logging this metadata.

This version also has more providers enabled. This will
allows us to remove the code in PBnJ for these providers.
For example, redfish.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>