Fix orphaned goroutine in worker:
The goroutine (in worker) dequeuing tasks was a
blocking call. This meant that after there were
no tasks in the queue, this dequeue call would sit
waiting. This meant that when the rest of the worker
exited this goroutine did not and was orphaned. Workers
are per host id so depending on unique hosts that
requested tasks this would grown and stay at that number.
The resolution was to make the dequeue call non-blocking.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>