Add support for sending NMI
Add SendNMI method to the gRPC diagnostic service. The method has
also been added to the client.

The version of bmclib is updated to use the new SendNMI code.

All of the protobuf go files seem to have changed due to minior
formatting changes and comments from a newver version of protoc.

Signed-off-by: John Mears <jmears@equinix.com>